+++ CHANGELOG.md
@@ -9,10 +9,12 @@ Exclamation symbols (:exclamation:) note something of importance e.g. breaking c
 ### Added
 - Bot API 5.2 (Payments 2.0).
 ### Changed
+- `Telegram::runCommands` returns array of `ServerResponse` objects of executed commands.
 ### Deprecated
 ### Removed
 ### Fixed
 - Regex for namespace extraction from custom command classes.
+- Nested and user-triggered `Telegram::runCommands`.
 ### Security
 
 ## [0.72.0] - 2021-04-16